+++ frontend/src/app/App.tsx
@@ -1,16 +1,20 @@
-import React from 'react';
-import { Provider } from 'react-redux';
-import LandingPage from '../pages/landing';
-import LoggedPage from '../pages/logged';
-import configureStore from '../store';
+import React from "react";
+import { ConnectedRouter } from "connected-react-router";
+import { Provider } from "react-redux";
+import { BrowserRouter } from "react-router-dom";
+import configureStore, { history } from "../store";
+import Router from "./router";
 
 const store = configureStore({});
 
 const App = () => (
-  <Provider store={store}>
-    {/* <LandingPage/> */}
-    <LoggedPage/>
-  </Provider>
+  <BrowserRouter>
+    <Provider store={store}>
+      <ConnectedRouter history={history}>
+        <Router />
+      </ConnectedRouter>
+    </Provider>
+  </BrowserRouter>
 );
 
 export default App;