+++ frontend/src/index.tsx
@@ -3,5 +3,4 @@ import ReactDOM from "react-dom";
 import App from "./app/App";
 import "../sass/style.scss";
 
-
 ReactDOM.render(<App />, document.getElementById("react-root-index"));