+++ frontend/src/pages/logged/index.tsx
@@ -8,7 +8,6 @@ import BudgetTabs from "./budget-container/components/budget-tabs";
 
 const LoggedPage = () => {
   const dispatch = useDispatch();
-  const [loginForm, setLoginForm] = React.useState(true);
   const restCheck = useSelector((state: RootStateOrAny) => state.restCheck);
   useEffect(() => {
     const action = creators.fetchRestCheck();
@@ -20,14 +19,7 @@ const LoggedPage = () => {
       <AbsoluteBackground />
       {/* <Container fluid="md" className="m-5 mx-auto align-middle"> */}
       <Container fluid="md" className="logged-page mx-auto align-middle">
-
-        <BudgetContainer/>
-
-          
-
-          {/* <h1 className="title">Family Budget</h1>
-          <h3 className="subtitle">Gather your spendings.</h3>
-          <div>{restCheck.result}</div> */}
+        <BudgetContainer />
       </Container>
     </>
   );