+++ docs/CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.3.2]] - 2022-03-24
+
+###### _([diff: 0.3.1-0.3.2])_
+
+### Added
+
+- Routing ([`Router.tsx`](../frontend/src/app/Router.tsx))
+    - Budgets View for `/budgets`
+    - Landing page for any other URL
+
+### Fixed
+
+- Source Map warnings
+
 ## [[0.3.1]] - 2022-03-24
 
 ###### _([diff: 0.3.0-0.3.1])_
@@ -139,6 +153,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.3.1-0.3.2]: https://github.com/Luzkan/FamilyBudget/compare/0.3.1...0.3.2
 [diff: 0.3.0-0.3.1]: https://github.com/Luzkan/FamilyBudget/compare/0.3.0...0.3.1
 [diff: 0.2.2-0.3.0]: https://github.com/Luzkan/FamilyBudget/compare/0.2.2...0.3.0
 [diff: 0.2.1-0.2.2]: https://github.com/Luzkan/FamilyBudget/compare/0.2.1...0.2.2
@@ -146,6 +161,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.3.2]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.2
 [0.3.1]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.1
 [0.3.0]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.0
 [0.2.2]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.2.2